<==== Mercurial data ====>
Branch: master

+++ mediaserver/src/rest/handlers/image_handler.cpp
@@ -39,7 +39,7 @@ QnCompressedVideoDataPtr getNextArchiveVideoPacket(QnServerArchiveDelegate& serv
     }
 
     // if ceilTime specified try frame with time > requested time (round time to ceil)
-    if (ceilTime != AV_NOPTS_VALUE && video && video->timestamp < ceilTime - 1000ll)
+    if (ceilTime != (qint64)AV_NOPTS_VALUE && video && video->timestamp < ceilTime - 1000ll)
     {
         for (int i = 0; i < MAX_GOP_LEN; ++i) 
         {
